bottom sheet cleanup 2
